+++ src/backend/core/api/viewsets.py
@@ -741,10 +741,9 @@ class DocumentAccessViewSet(
     def perform_create(self, serializer):
         """Add a new access to the document and send an email to the new added user."""
         access = serializer.save()
-        language = self.request.headers.get("Content-Language", "en-us")
 
         access.document.email_invitation(
-            language,
+            access.user.language,
             access.user.email,
             access.role,
             self.request.user,
@@ -984,10 +983,11 @@ class InvitationViewset(
         """Save invitation to a document then send an email to the invited user."""
         invitation = serializer.save()
 
-        language = self.request.headers.get("Content-Language", "en-us")
-
         invitation.document.email_invitation(
-            language, invitation.email, invitation.role, self.request.user
+            self.request.user.language,
+            invitation.email,
+            invitation.role,
+            self.request.user,
         )