+++ src/backend/core/tests/documents/test_api_document_accesses_create.py
@@ -16,6 +16,9 @@ from core.tests.conftest import TEAM, USER, VIA
 pytestmark = pytest.mark.django_db
 
 
+# Create
+
+
 def test_api_document_accesses_create_anonymous():
     """Anonymous users should not be allowed to create document accesses."""
     document = factories.DocumentFactory()
@@ -123,7 +126,7 @@ def test_api_document_accesses_create_authenticated_administrator(via, mock_user
             document=document, team="lasuite", role="administrator"
         )
 
-    other_user = factories.UserFactory()
+    other_user = factories.UserFactory(language="en-us")
 
     # It should not be allowed to create an owner access
     response = client.post(
@@ -198,7 +201,7 @@ def test_api_document_accesses_create_authenticated_owner(via, mock_user_teams):
             document=document, team="lasuite", role="owner"
         )
 
-    other_user = factories.UserFactory()
+    other_user = factories.UserFactory(language="en-us")
 
     role = random.choice([role[0] for role in models.RoleChoices.choices])
 
@@ -233,3 +236,72 @@ def test_api_document_accesses_create_authenticated_owner(via, mock_user_teams):
         in email_content
     )
     assert "docs/" + str(document.id) + "/" in email_content
+
+
+@pytest.mark.parametrize("via", VIA)
+def test_api_document_accesses_create_email_in_receivers_language(via, mock_user_teams):
+    """
+    The email sent to the accesses to notify them of the adding, should be in their language.
+    """
+    user = factories.UserFactory()
+
+    client = APIClient()
+    client.force_login(user)
+
+    document = factories.DocumentFactory()
+    if via == USER:
+        factories.UserDocumentAccessFactory(document=document, user=user, role="owner")
+    elif via == TEAM:
+        mock_user_teams.return_value = ["lasuite", "unknown"]
+        factories.TeamDocumentAccessFactory(
+            document=document, team="lasuite", role="owner"
+        )
+
+    role = random.choice([role[0] for role in models.RoleChoices.choices])
+
+    assert len(mail.outbox) == 0
+
+    other_users = (
+        factories.UserFactory(language="en-us"),
+        factories.UserFactory(language="fr-fr"),
+    )
+
+    for index, other_user in enumerate(other_users):
+        expected_language = other_user.language
+        response = client.post(
+            f"/api/v1.0/documents/{document.id!s}/accesses/",
+            {
+                "user_id": str(other_user.id),
+                "role": role,
+            },
+            format="json",
+        )
+
+        assert response.status_code == 201
+        assert models.DocumentAccess.objects.filter(user=other_user).count() == 1
+        new_document_access = models.DocumentAccess.objects.filter(
+            user=other_user
+        ).get()
+        other_user_data = serializers.UserSerializer(instance=other_user).data
+        assert response.json() == {
+            "id": str(new_document_access.id),
+            "user": other_user_data,
+            "team": "",
+            "role": role,
+            "abilities": new_document_access.get_abilities(user),
+        }
+        assert len(mail.outbox) == index + 1
+        email = mail.outbox[index]
+        assert email.to == [other_user_data["email"]]
+        email_content = " ".join(email.body.split())
+        if expected_language == "en-us":
+            assert (
+                f"{user.full_name} shared a document with you: {document.title}"
+                in email_content
+            )
+        elif expected_language == "fr-fr":
+            assert (
+                f"{user.full_name} a partagé un document avec vous: {document.title}"
+                in email_content
+            )
+        assert "docs/" + str(document.id) + "/" in email_content